added route information, created deb package, refactored install and uninstall scripts

+++ install.sh
@@ -5,23 +5,50 @@ if [ "$(id -u)" != "0" ]; then
   exit 1
 fi
 
+echo "Preparing to compile C++ source"
 mkdir build
 cd build
 
-cmake ../src/cpp/
-make
+echo "Building Makefile"
+cmake ../src/cpp/ &> /dev/null
 
+if [ "$?" != "0" ]; then
+  echo "Error while building Makefile. Exiting"
+  exit 1
+fi
+
+echo "Compiling..."
+make &> /dev/null
+if [ "$?" != "0" ]; then
+  echo "Error while compiling C++ source. Exiting"
+  exit 1
+fi
+
+echo "Installing C++ shared object "
 mkdir -p /usr/local/share/lonet
 cp libnet_if_binding.so /usr/local/share/lonet
 
 cd ..
 rm -rf ./build
 
+echo "Creating Python3 virtual environment"
 python3 -m venv venv
 source venv/bin/activate
 
-python3 -m pip install pyinstaller
-pyinstaller --onefile ./src/main.py --name lonet
+echo "Installing pyinstaller dependency into venv"
+python3 -m pip install pyinstaller &> /dev/null
+
+if [ "$?" != "0" ]; then
+  echo "Cannot install pyinstaller. Exiting"
+  exit 1
+fi
+
+echo "Compiling python source..."
+pyinstaller --onefile ./src/main.py --name lonet &> /dev/null
+if [ "$?" != "0" ]; then
+  echo "Error while compiling python source. Exiting"
+  exit 1
+fi
 
 deactivate
 mkdir -p bin
@@ -29,4 +56,5 @@ mkdir -p bin
 cp ./dist/lonet ./bin
 rm -rf ./dist ./build ./lonet.spec ./venv
 
+echo "Installing compiled binary"
 cp ./bin/lonet /usr/local/bin/lonet
